+++ hata/discord/stage/tests/test__Stage__utility.py
@@ -14,12 +14,10 @@ def test__Stage__copy():
     """
     Tests whether ``Stage.copy`` works as intended.
     """
-    channel_id = 202303110043
     privacy_level = PrivacyLevel.public
     topic = 'breaking me softly'
     
     stage = Stage(
-        channel_id = channel_id,
         privacy_level = privacy_level,
         topic = topic,
     )
@@ -37,12 +35,10 @@ def test__Stage__copy_with__0():
     
     Case: No fields given.
     """
-    channel_id = 202303110044
     privacy_level = PrivacyLevel.public
     topic = 'breaking me softly'
     
     stage = Stage(
-        channel_id = channel_id,
         privacy_level = privacy_level,
         topic = topic,
     )
@@ -60,29 +56,24 @@ def test__Stage__copy_with__1():
     
     Case: No fields given.
     """
-    old_channel_id = 202303110045
     old_privacy_level = PrivacyLevel.public
     old_topic = 'breaking me softly'
     
-    new_channel_id = 202303110046
     new_privacy_level = PrivacyLevel.guild_only
     new_topic = 'breaking me softly'
     
     stage = Stage(
-        channel_id = old_channel_id,
         privacy_level = old_privacy_level,
         topic = old_topic,
     )
     
     copy = stage.copy_with(
-        channel_id = new_channel_id,
         privacy_level = new_privacy_level,
         topic = new_topic,
     )
     _assert_fields_set(copy)
     vampytest.assert_is_not(copy, stage)
     
-    vampytest.assert_eq(copy.channel_id, new_channel_id)
     vampytest.assert_is(copy.privacy_level, new_privacy_level)
     vampytest.assert_eq(copy.topic, new_topic)
 
@@ -185,11 +176,10 @@ def test__Stage__channel():
     """
     channel_id = 202303110054
     
-    for input_value, expected_output in (
-        (0, None),
-        (channel_id, Channel.precreate(channel_id)),
+    for stage_id, input_value, expected_output in (
+        (202303110060, channel_id, Channel.precreate(channel_id)),
     ):
-        stage = Stage(channel_id = input_value)
+        stage = Stage.precreate(stage_id, channel_id = input_value)
         vampytest.assert_is(stage.channel, expected_output)